requirements.txt
python-telegram-bot
pytz

